+++ nifti2/nifti_tool.c
@@ -922,7 +922,7 @@ int add_int(int_list * ilist, int val)
 {
    if( ilist->len == 0 ) ilist->list = NULL;  /* just to be safe */
    ilist->len++;
-   ilist->list = (int *)realloc(ilist->list,ilist->len*sizeof(int));
+   ilist->list = (int *)realloc(ilist->list, (size_t)ilist->len * sizeof(int));
    if( ! ilist->list ){
       fprintf(stderr,"** failed to alloc %d (int *) elements\n",ilist->len);
       return -1;
@@ -943,7 +943,7 @@ int add_string(str_list * slist, const char * str)
 {
    if( slist->len == 0 ) slist->list = NULL;  /* just to be safe */
    slist->len++;
-   slist->list = (const char **)realloc(slist->list,slist->len*sizeof(char *));
+   slist->list = (const char **)realloc(slist->list, (size_t)slist->len * sizeof(char *));
    if( ! slist->list ){
       fprintf(stderr,"** failed to alloc %d (char *) elements\n",slist->len);
       return -1;
@@ -2358,14 +2358,14 @@ static char * read_file_text(const char * filename, int * length)
 
    /* allocate the bytes, and fill them with the file contents */
 
-   text = (char *)malloc(len64 * sizeof(char));
+   text = (char *)malloc((size_t)len64 * sizeof(char));
    if( !text ) {
       fprintf(stderr,"** RFT: failed to allocate %" PRId64 " bytes\n", len64);
       fclose(fp);
       return NULL;
    }
 
-   bytes = fread(text, sizeof(char), len64, fp);
+   bytes = fread(text, sizeof(char), (size_t)len64, fp);
    fclose(fp); /* in any case */
 
    if( bytes != (size_t)len64 ) {
@@ -2562,7 +2562,7 @@ int remove_ext_list( nifti_image * nim, const char ** elist, int len )
    if( g_debug > 2 )
       fprintf(stderr,"+d removing %d exts from '%s'\n", len, nim->fname );
 
-   if( ! (marks = (int *)calloc(nim->num_ext, sizeof(int))) ) {
+   if( ! (marks = (int *)calloc((size_t)(nim->num_ext),sizeof(int))) ) {
       fprintf(stderr,"** failed to alloc %d marks\n",nim->num_ext);
       return -1;
    }
@@ -3929,7 +3929,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                }
                /* otherwise, we're good */
                { const int16_t sval = (int16_t)val;
-                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(sval), &sval,(size_t)sizeof(sval)); }
+                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(sval), &sval,sizeof(sval)); }
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %d\n",
                           fc, field->name, val);
@@ -3949,7 +3949,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                   return 1;
                }
                { const int32_t ival = (int32_t)val;
-                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(ival), &ival,(size_t)sizeof(ival)); }
+                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(ival), &ival,sizeof(ival)); }
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %d\n",
                           fc, field->name, val);
@@ -3969,7 +3969,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                           fc,field->len);
                   return 1;
                }
-               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(v64), &v64,(size_t)sizeof(v64));
+               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(v64), &v64,sizeof(v64));
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %" PRId64 "\n",
                           fc, field->name, v64);
@@ -3989,7 +3989,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                   return 1;
                }
                /* otherwise, we're good */
-               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(fval), &fval,(size_t)sizeof(fval));
+               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(fval), &fval,sizeof(fval));
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %f\n",
                           fc, field->name, fval);
@@ -4010,7 +4010,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                   return 1;
                }
                /* otherwise, we're good */
-               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(f64), &f64,(size_t)sizeof(f64));
+               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(f64), &f64,sizeof(f64));
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %f\n",
                           fc, field->name, f64);
@@ -4022,10 +4022,10 @@ int modify_field(void * basep, field_s * field, const char * data)
          case NT_DT_STRING:
          {
             char * dest = (char *)basep + field->offset;
-            nchars = dataLength;
-            strncpy(dest, data, field->len);
+            nchars = (int)dataLength;
+            strncpy(dest, data, (size_t)(field->len));
             if( nchars < field->len )  /* clear the rest */
-               memset(dest+nchars, '\0', field->len-nchars);
+               memset(dest+nchars, '\0', (size_t)(field->len-nchars));
          }
          break;
    }
@@ -4188,7 +4188,7 @@ static int convert_NBL_data(nifti_brick_list * NBL, int old_type, int new_type,
    nifti_datatype_sizes(new_type, &nbyper, NULL);
    NBLnew.bsize = nbvals * nbyper;
    NBLnew.nbricks = NBL->nbricks;
-   NBLnew.bricks = (void **)calloc(NBLnew.nbricks, sizeof(void *));
+   NBLnew.bricks = (void **)calloc((size_t)NBLnew.nbricks, (size_t)(sizeof(void *)));
    if( ! NBLnew.bricks ) {
       fprintf(stderr,"** cNBLd: failed to allocate %" PRId64 " void pointers\n",
               NBLnew.nbricks);
@@ -4280,7 +4280,7 @@ static int convert_raw_data(void ** retdata, void * olddata, int old_type,
 
    /* allocate new memory (calloc, in case of partial filling) */
    nifti_datatype_sizes(new_type, &nbyper, NULL);   /* get nbyper */
-   newdata = calloc(nvox, nbyper);
+   newdata = calloc((size_t)nvox, (size_t)nbyper);
    if( !newdata ) {
       fprintf(stderr,"** failed to alloc for %" PRId64 " %s elements\n",
               nvox, typestr);
@@ -7638,7 +7638,7 @@ nifti_image * nt_read_bricks(nt_opts * opts, char * fname, int len,
     /* now populate NBL (can be based only on len and nim) */
     NBL->nbricks = len;
     NBL->bsize = nim->nbyper * nim->nx * nim->ny * nim->nz;
-    NBL->bricks = (void **)calloc(NBL->nbricks, sizeof(void *));
+    NBL->bricks = (void **)calloc((size_t)(NBL->nbricks), (size_t)(sizeof(void *)));
     if( !NBL->bricks ){
         fprintf(stderr,"** NRB: failed to alloc %" PRId64 " pointers\n",
                 NBL->nbricks);
@@ -7653,7 +7653,7 @@ nifti_image * nt_read_bricks(nt_opts * opts, char * fname, int len,
 
     /* now allocate the data pointers */
     for( c = 0; c < len; c++ ) {
-        NBL->bricks[c] = calloc(1, NBL->bsize);
+        NBL->bricks[c] = calloc(1, (size_t)(NBL->bsize));
         if( !NBL->bricks[c] ){
             fprintf(stderr,
                     "** NRB: failed to alloc brick %d of %" PRId64 " bytes\n",
